cue: add build options
Add a way for the user to specify build options (passed to compile
and related functions). These are encoded as Python data classes.
Also implement the conversion to the C representation expected by
`libcue`.

Because `Scope` references a `Value`, also add a stub for `Value`.

This functionality will be tested in a subsequent CL.

Updates #3100.
Updates #3073.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>
Change-Id: Idae52d9cfdf2121009c03127f7b2d0b12591a495