libcue: pass full libcue binary name to dlopen
On macOS dlopen(3) can take the base name of a library, such as
`cue`, and look for `libcue.dylib`.

On Linux this is not always the case. At least with GNU libc,
dlopen(3) expects a name like `libcue.so`.

On Windows LoadLibrary can use the base name without `.dll`, but
for consistency use the full file name `cue.dll` on Windows as well.

Also add Linux and Windows to CI.

Fixes cue-lang/cue#3190.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>
Change-Id: I52d3418ee39477be8e48ff7af7366e593a313fcb