.gitattributes: normalize file endings to LF
Normalize file endings to LF, this is to silence git status warnings
on Windows, which in turn brings us closer to enabling the CI on Windows.

CI on Windows will only be enabled in a subsequent CL because it
requires further configuration.

Updates cue-lang/cue#3253.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>
Change-Id: I97619c2c41d0572b6d3eb215a7bd8091d002dd66
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue-py/+/1196827
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Roger Peppe <rogpeppe@gmail.com>