internal/ci: add libcue to PATH on Windows
Windows looks for DLLs in PATH, so add libcue to PATH.

GitHub actions don't allow setting PATH via `env`, rather we need
to append to the file pointed to by `$GITHUB_PATH`. Do this in a
separate step because this will only affect future steps.

This brings us one step closer to enabling CI on Windows, but we
don't do it in this CL yet as it requires additional work.

Updates cue-lang/cue#3253.

Signed-off-by: Aram Hăvărneanu <aram@mgk.ro>
Change-Id: I3f7e5d8acfed51961c3eaefd727ce76c3e08ae70