internal/ci: use latest stable Go in CI
Now that we have moved past the min go1.22.x requirement, we can rely on
the latest stable version, which is generally what the end user will be
doing.

We leave the python version untouched for now, although this should
likely "move" like latest for the same reason.

Change-Id: I4d8ca1a45463671cd2a9b473cbc868ea3c6dac35
Signed-off-by: Paul Jolly <paul@myitcv.io>
Reviewed-on: https://review.gerrithub.io/c/cue-lang/cue-py/+/1216796
TryBot-Result: CUEcueckoo <cueckoo@cuelang.org>
Reviewed-by: Daniel Martí <mvdan@mvdan.cc>